Controller : Statistic
- Auto-génération du Controller et twig index de Statistic,
- Ajout de route pour voir chaque type de Statistique, et création de leur twig
- Ajout d'un fragment.
- Modification de l'entité pour que "value" soit un int
- Génération de 10 Statistiques par restaurant dans la fixture.
- Traduction en français de mots dans la navbar

+++ src/Entity/Statistic.php
@@ -17,7 +17,7 @@ class Statistic
     private ?string $statisticType = null;
 
     #[ORM\Column(nullable: true)]
-    private ?float $value = null;
+    private ?int $value = null;
 
     #[ORM\Column(nullable: true)]
     private ?\DateTime $date = null;
@@ -43,12 +43,12 @@ class Statistic
         return $this;
     }
 
-    public function getValue(): ?float
+    public function getValue(): ?int
     {
         return $this->value;
     }
 
-    public function setValue(?float $value): static
+    public function setValue(?int $value): static
     {
         $this->value = $value;
